+++ src/helpers/StringHelper.php
@@ -433,8 +433,8 @@ class StringHelper extends BaseStringHelper
      * @param string $template The template to parse. The template may contain double curly brackets variables.
      * @param array $variables The variables which should be available in the template.
      * @param boolean $removeEmpty Whether variables in double curly brackets should be removed, even the have not be assigned by $variables array.
-     * @param string $leftDelimiter The delimiter for the variable on the left, default is `{{`
-     * @param string $rightDelimiter The delimiter for the variable on the right, default is `}}`
+     * @param string $leftDelimiter The delimiter for the variable on the left, default is `{{` {@since 1.2.0}
+     * @param string $rightDelimiter The delimiter for the variable on the right, default is `}}` {@since 1.2.0}
      * @return string
      */
     public static function template($template, array $variables = [], $removeEmpty = false, $leftDelimiter = '{{', $rightDelimiter = '}}')