+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file. This projec
 
 ## 1.2.0 (15. June 2021)
 
-+ [#]() Added option to define the delimiter in `StringHelper::template` function.
++ [#4](https://github.com/luyadev/yii-helpers/pull/4) Added option to define the delimiter in `StringHelper::template` function.
 
 ## 1.1.1 (6. April 2021)