+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this project will be documented in this file. This project adheres to [Semantic Versioning](http://semver.org/).
 
+## 1.2.0 (15. June 2021)
+
++ [#4](https://github.com/luyadev/yii-helpers/pull/4) Added option to define the delimiter in `StringHelper::template` function.
+
 ## 1.1.1 (6. April 2021)
 
 + [#2](https://github.com/luyadev/yii-helpers/issues/2) Fix issue where the highlight text has been highlight to.